Remove duplicate class analysis from memory
We no longer store the class analysis of a compile task's
output in that compile tasks's own cache. Instead, we compute
it right before we need it when the next compilation starts.
This removes a lot of unnecessary memory pressure, because no
one else was reusing that state.

It would be even better if we could reuse the analysis that
downstream tasks did on that classes directory (which is an input
for them). To do that we'd need to ensure consistent hashing strategies
between the two.