Recompile types in a package if its package-info changes
Annotation processors can inspect the package of a class,
so we need to make sure that that information is up-to-date
when the package-info changes.

To support this feature we now remember all classes in the
class analysis, not just the ones that had dependents. This
leads to a small increase in memory usage, but means we can
accurately recompile only the classes affected by the given
package change. Alternatively we could omit that data and do
a full recompile instead.