Intern Strings when read from the backing store of the file hasher
We were interning the strings when putting them into the in-memory cache,
but when that cache got full it would start discarding entries and later
loading them from the backing store. The load operation would return a
non-interned String, resulting in much higher memory pressure. This is
now fixed by using an interning serializer.