Intern class names when reading back from the cache
We intern class names when we analyze classes, but until
now we did not intern them when reading back from the on-disk
cache. This meant that builds with pre-existing caches would
have much higher memory footprint, just like builds that exceed
the in-memory size limit. With this change all builds now have
the same low footprint.