Separate 'prefer' and 'require' in dependency versions
When we introduced the ability to declare a 'preferred' version on
a dependency declaration, this was implemented such that declaring
a "required"  dependency version using `org:foo:1.0` was effectively
the same as declaring a "preferred" version `org:foo { prefer '1.0' }`.

In order to differentiate between the behaviour of required and
preferred dependency versions, this commit introduces a separate
model for these constraint types. This model is published to
Gradle `.module` metadata files, and is retained internally
throughout dependency resolution.

At this stage, the behaviour of required and preferred versions
is identical. A later commit will introduce the behavioural
difference.