Clean up distributions at most every 24 hours
Since the distribution cleanup was made independent of the
version-specific cache cleanup in #5981, it was run every time the
daemon was stopped. Now, instead of introducing yet another
`gc.properties` file, it is only executed if version-specific cache
cleanup was executed but regardless whether the latter actually deleted
something.

Fixes #6083.