Reinstate requirement for removal details in deprecation notices (#6080)
* Reinstate requirement for removal details in deprecation notices

If something is deprecated, it must be removed at some point (otherwise there is no point). Removing doesn't have to mean “physically” removing the _thing_. In the case in question, we will be removing the functionality but keeping the method (because its part of an interface we have to fulfill).