Improve file lock contention handling
This commit improves the likelihood for lock requesters to acquire a
file lock after it is released due to contention. After the lock has
been released, the former lock holder now sends a packet to the sockets
of all requesters. While old clients will simply ignore the additional
packet, new clients will interpret it as a signal that the file lock has
been released and will try to acquire it immediately.

Issue: gradle/gradle-private#1412.