Reduce need for contention during cache cleanup
Cache cleanup now creates an immutable copy of the file access journal
when checking the first file and bulk deletes access times for removed
files after the cleanup is done. Thus, the file lock for the journal is
only needed during the time it takes to create the copy and when bulk
deleting access times but not for reading access times during the
cleanup operation.

Fixes gradle/gradle-private#1412.