Add required repositories block to samples
Starting with Kotlin 1.2.60.*, afaict, the Kotlin plugin downloads the
Kotlin compiler and its dependencies on demand using a project
configuration, which requires a repository to be configured. These
repositories were not necessary until the upgrade to Kotlin 1.2.60
because the kotlin-eap repository was being automatically added to
every project (to ease migration).