Add `MutableVersionConstraint.require(version)`
By default, a dependency like 'org:foo:1.0' defines a 'require' version
constraint.

With this change, it is also possible to set the 'required' version for
a dependency from within `Configuration.withDependencies` or a component
metadata rule. This brings 'require' on par with 'prefer' and 'strictly'
for the purposes of configuration.