Introduce the `existing` property delegate
To allow provider based access to collection elements via delegated
properties.

In contrast to the value based access currently allowed by the
default property delegate:

```kotlin
val clean by tasks
assertInferredTypeOf(
    clean,
    typeOf<Task>() // default delegate means value access
)
```

With the `existing` delegate, the type of the property will be
inferred to be the same as the provider type specialized to the
container:

```kotlin
val clean by tasks.existing
assertInferredTypeOf(
    clean,
    typeOf<TaskProvider<Task>>()
)

val main by sourceSets
assertInferredTypeOf(
    main,
    typeOf<NamedDomainObjectProvider<SourceSet>>()
)
```

In order to make the distinction between `TaskProvider` and
`NamedDomainObjectProvider` possible, this commit also introduces the
`TaskContainerScope` type and a few other wrapper types that help
carry the necessary types through the type inference process.

As a bonus, a test demonstrates how the `existing` delegate can be used
as the hook for well-known element accessors.