Handle missing type correctly
When communicating with an old version, the length of the received
packet will be shorter than the byte array. Now, `FileLockPacketPayload`
only reads the type when it has been received and returns UNKNOWN if it
has not. Previously, it would have worked by accident because the last
byte was zero which corresponds to the ordinal number of UNKNOWN.