Rename private fields to work around Groovy compiler bug
A recently discovered Groovy compiler bug (see https://issues.apache.org/jira/browse/GROOVY-8732) causes subclasses of AbstractTask to refer to internal types when the name of a property on AbstractTask matches the name of a private field. To work around this problem until the Groovy compiler is fixed the private fields are renamed.