Only run static Groovy compile subclassing tests with Gradle 2.1+
Groovy introduced static compilation ({@link groovy.transform.CompileStatic}) in Groovy 2.0.0. We switched to using Groovy 2.3.3 from 1.8.6 in Gradle 2.0. However, Groovy 2.3.3 shipped with Gradle 2.0 had a bug that prevents the test to be compiled. Thus the first version we test with is Gradle 2.1 that shipped with Groovy 2.3.6 which fixed that issue.