Show conflict resolution reason for 2nd level dependencies
The dependencyInsight report shows a full suite of selection reasons
for the dependency being analysed, but shows only the selected version
for dependencies that are transitively responsible for the target
dependency.

With this change, we report the conflict resolution reason for any higher-order
dependencies that were selected through conflict resolution.