Add test showing different resolution based on locking being enabled
The current implementations of `LockingAwareSelector` result in different
resolution results being produced for a configuration based on whether
locking has been enabled for that configuration. This occurs in the absense
of any locking constraints, due to the way we change the interpretation of
`+` and `latest.` selectors when locking is merely enabled.
`