Support for gradle wrapper distribution mirror
I first [opened a discussion on the Gradle forums][discuss] back in May
6th, 2018.  I hadn't seen any activity one way or another.

I feel this is a very relevant feature for any company who desires to use
gradle in significant capacity.  It makes sense for a company to mirror
sources of download for all of their project dependencies.  Gradle
through Gradle wrapper is no different from that desirable perspective.

Hopefully, this is up to snuff for the Gradle devs.  Nobody from the
development team responded to my original forum post before I did this
work.

[discuss]: https://discuss.gradle.org/t/support-for-gradle-wrapper-distribution-mirror/26838

I agree to the Developer Certificate of Origin.

Signed-off-by: Sam Gleske <sam.mxracer@gmail.com>