Use compileAll to warm up the build cache
Sanity Check can run as a job in parallel to the functional tests.
This allows getting faster feedback for integration tests and
makes the common initial job faster.

The build cache should be populated with the whole compilation result
by this initial job, so the follow up jobs should continue using the
result of compilation via the build cache.

Fixes gradle-private#1440