Make watchpoint registration faster
The watchpoint registration algorithm was quadratic in the
number of input files that all the tasks in a build had.

This was particularly visible in the case of the Kotlin plugin,
which had one task that used individual class files instead of a
file tree for a class directory. The overhead for that project
dropped from about 1 minute to near zero.

See https://github.com/gradle/gradle/issues/5939