Make scope mappings of POM live
The scope mappings of the POM file of the Upload task
were a copy of the mappings defined in the project.

This meant that depending on buildscript ordering mappings
might be lost if the POM was configured before all mappings
had been added. This is now fixed by making the mappings live.