Make "belongsTo" create virtual components by default
This commit changes the behavior of "belongsTo" in component metadata rules,
so that it creates a virtual platform by default. By doing this, we optimize
the engine which can now avoid checking for a real, published platform,
avoiding costly repository calls to fetch metadata.

If a component needs to declare that it participates in a published platform,
or if we don't know if the platform is published or not, then the alternate,
new `belongsTo(id, virtual)` method needs to be called.

This is a performance optimization only.

Closes #6447