Recompile when a resource file changes
We can't tell what annotation processors might be doing with a
resource file, as the processing API does not provide any way
of associating resources with outputs. Therefor the only safe
option for now is doing a full recompile.

We no longer force a full recompile when a processor reads a
resource. We only recompile if the resource changes. This makes
it even easier to make a lot of processors incremental.