Add support for enforcing published platforms
This commit aligns the behavior of Gradle for published platforms
to the one for virtual platforms with regards to enforced behavior.

The implementation is pretty weak as it uses a component metadata
rule to update the dependencies/constraints of a platform, which
implies that the version used in the DSL is a fixed version.