Use our own lock for SystemProperties methods
We attempted to coordinate all access to system properties
for maximum security, but that didn't work out because it
also prevented reads from other threads while someone was
holding the lock. This is no longer a problem on Java 9,
but a deadlock is possible on Java 8.

We now only coordinate our own writes and accept that some
3rd party code may use `setProperty()` and bypass our checks
that way.