Ensure plugin bootstrapping in composites works without workarounds
When a plugin is built inside a composite build and that plugin uses
and older version of itself to build itself, it should not be substituted
into its own `buildscript {}` classpath. This only used to work with some
workarounds, making composites unnecessarily hard to use in that case.
Since Gradle 4.9 this works out of the box. This test change just makes
sure we don't regress again like we did in Gradle 4.1.