Rebaseline cold daemon performance test
This test has crept up a tiny bit over time so that there is possibly a tiny
regression (up to ~50ms).  There is no clear set of changes that caused this
regression and the test almost always passes.  There have been a number of
retries over recent weeks and one "failure" where we were unlucky to
have several retries fail in a row.  Given the small regression and the
lack of a smoking gun with respect to cause, it's not worth the time
required to investigate/fix.  Rebaselining to eliminate flakiness.