Fix handling of 'prefer' constraints
This commit fixes a bug that resulted in us giving too much weight
to 'prefer' constraints when combined with other constraints.
By resolving each prefer constraint and trying each result in turn
starting with the highest, the processing of prefer constraints is
now predictable and reasonable.

This work highlights the need for more work to clean up the way that
we resolve multiple selectors for a module. Instead of resolving each
selector independently and then attempting to merge the results, we
should be merging the selectors first (where possible), prior to
resolving. This would reduce unnecessary dynamic resolves in the
case where we have both a range constraint `[10, 12]` and a fixed
version constraint `11`. A good example of where this would benefit
would be dependency locking, where all version ranges are effectively
constrained to a single version.