Exclude Ant dependency of cglib
Spock 1.2-RC2 bumped its cglib dependency to 3.7.2 which changed its
Ant dependency to 1.10.3 which is incompatible with the current version
constraint that strictly requires 1.9.11. Since we cannot yet upgrade
to Ant 1.10.x because it requires Java 8, we now use a component
metadata rule to remove the dependency from cglib to Ant.