Remove support for optional dependencies as constraints
Whenever we used the "improved POM support" experimental flag, Gradle was
interpreting optional dependencies as constraints. It meant that if a
hard dependency was found in the graph, and that another module had an
optional dependency on it, then we would conflict resolve between the two
versions.

It appears that this behavior causes some troubles, for example mentioned
in #5434. In Maven, optional dependencies are really never used. They are
pure documentation.

We also want to provide a better alternative to optional dependencies,
which will include a better model of optional features. Support for optional
dependencies coming from a Maven POM may be included as part of this work.