Optimize attribute matching
Attribute coercion is a pretty expensive process. But we very often coerce the same
attributes to the same other attribute types. The consequence is that we do the same
amount of work repeatedly. Knowing that attributes are interned thanks to the
attribute factory, and that their lifetime is bound to a single build, we can use
a coercion cache that uses a simple map, to make it faster. The map, however, needs
to be a concurrent map because we have no guarantee that this method cannot be
called concurrently.

In addition, the multiple candidate matcher is now aware that it's pretty usual
to find the same attributes everywhere, so it optimizes for this case when collecting
the candidate values, by avoiding the creation of a hash set.