Fix skipped performance scenarios
`AbstractGradleBuildPerformanceTestRunner` uses `org.junit.Assume` to
throw an `AssumptionViolatedException` for scenarios that should be
skipped. Spock 1.1 changed how `AssumptionViolatedExceptions` are
handled when thrown from code in `then` blocks. While in Spock 1.0 the
test was reported as skipped, it is reported as failed in Spock >= 1.1.
Thus, affected performance tests now call `run()` in a `when` block
and check for a non-null results object in a `then` block.