Fix Gradleception build
Gradle 5.0 changes the behavior of BOM import, which is now required to be
imported explicitly using the `platform` keyword. However, this is not yet
available on the Gradle version we use to build, so this commit introduces
a temporary fix so that we can use the same syntax on all tested versions
of Gradle.