Rework collection of performance tests results
This commit changes the way the coordinator collects performance test results.
Previously, we had a list of test ids, and iterated only once on the list. If
the result of a test wasn't available, then we'd wait one minute, and recheck.
In practice, it means there was no way to have an idea of the global completion,
because the first tests in the list can "block" reporting.

Now, we loop on the full list of tests, check if the result of each test is
available. If so, we collect the result, and if not, requeue until all tests
are completed. We keep the one minute loop between checks. This allows us to
provide better feedback by having a global advancement metric.