Fix test report index page
Prior to this commit the Classes tab of the test report index page
was generated with the link to the class as a sibling to the table cell
that should have contained it. The update of jsoup discovered this
issue and the table cell now encloses the link to the test class page.