Speed up worker API cancellation
Previously we cancelled work items, but only after they
had already been started. This meant that cancellation could
take up to numberOfItems*timeToInterruptCheck time.

We now eagerly cancel the work items, so that they don't even
start executing. This limits the delay to the time it takes
a single item to interrupt.