Change the lifecycle for included builds so that a given build is loaded and configured at most once, and a given task is executed at most once, during a Gradle invocation.
Previously, an included build could be discarded and its state recreated when tasks from that build were required at both configuration time and execution time of some other builds (possibly different builds for each). This was in order to avoid breaking changes to some public hooks, such as `taskGraph.whenReady()`. However this behaviour just adds too much complexity to justify. Instead, we'll look at replacements and possibly some minor breakages for these hooks.

This change fixes a number of edge cases in mixing plugin and production dependencies on included builds. There are a number of issues with the implementation which will be fixed in later commits.