Lazily compute derived dependencies
Whenever we compute derived variants, we need to create new `ConfigurationMetadata`. Before, we used to compute
the dependencies of those derived metadata immediately, even if in practice there were a lot of chances that
they would only be needed on the selected variant. The only thing that we need to compute eagerly is the transformed
attributes. However, because attributes can be mutated by component metadata rules, we cannot create a lazy
wrapper on top of `ComponentMetadata`, so this commit bakes in the performance improvement directly to the
default configuration metadata class, which is now lazily computing a derived view of the dependencies based
on a enum corresponding to the different possible cases.