Never do version range merging for 'prefer' selectors
When no 'require' version was set for a dependency, we were processing
a set of 'prefer' version constraints in the same way as a set of 'require'
version constraints: ie we were doing version range merging for these
'prefer' constraints. When any 'require' constraint was provided, we no
longer did this merging for 'prefer' constraints.

With this commit, we no longer do this special processing for 'prefer'
in the absence of 'require'. Prefer constraints are processed consistently,
choosing the highest preference that does not conflict with any require
constraint.