Reduce cost of deriving variants for Maven metadata
Deriving variant cost is dominated by computing the derived attributes. However,
we know that for Maven metadata, the input attributes are always the same, and
what we use to derive too. So this commit introduces a specialized attributes
factory which goal is to cache concatenation of attributes used in derived
variants.