Update diagrams in Java Plugin chapter
I have made several changes:

 * Split the large graph of dependency configurations into two.
 * Simplified the resulting graphs with a focus on the 'main' and 'test' source
   sets.
 * Added a legend to help understand the graphs.
 * Tried to make the diagram style consistent with the Java Library Plugin
   chapter.

The main problem with the original graph was that it had so many elements on it
with many, many interconnecting lines that it was very difficult to understand.
In addition, much of the information wasn't particularly helpful. I think this
approach of simply focusing on 'main' and 'test' configurations helps in the
areas that are more heavily used and that cause most confusion.

Finishes #5681.