Make `PreferJavaRuntime` choose JAVA_RUNTIME if its present
This commit changes the prefer java runtime disambiguation rule, so that
it's more tolerant of unexpected candidate values. Before this change, the
rule would prefer the runtime variant if, and only if:

- the consumer expressed no preference for `Usage` and
   - the candidate values are _exactly_ `JAVA_RUNTIME` or `JAVA_API` or
   - the candidate values were only of the above or missing

Now that we cannot have missing (`null`) candidate values, the rule above
could already have been simplified. However, this commit change it to an
even simpler rule:

- if the consumer expressed no preference and the candidate values contain
`JAVA_RUNTIME`, use it.

This should allow us to be more tolerant whenever new candidate values are
introduced by plugins. Without this change, any external dependency from
Maven would suddenly start failing if the type of the component was unexpected.
This is because we would use derived variants, and that we do it whatever
the type of component is (jar, pom, bom, aar, ...)

This change will therefore restore the previous behavior, which was to choose
the "runtime" variant of an unknown component type (for external Maven dependencies).