Allow disabling of cache clean by the end user
This behavior is not wanted by all users, and some users actually find it very helpful. This PR adds a property (org.gradle.disable.cache.cleanup) that can be set in the home folder gradle.properties file, and if this property is set to true, the cache deletion will be skipped.

Signed-off-by: Thad House <thadhouse1@gmail.com>