Gradle init will not use parent `gradle.properties`
Running `gradle init` has always ignored a settings file in a parent
directory, instead creating an isolated, nested build with it's own
settings file. However, if a settings.gradle file was found in a
parent directory, a `gradle.properties` file found in the same directory
was used to determine daemon compatibility. This causes issues with our
NoDaemon integration tests, since the `init` task was finding the
properties file in the root Gradle directory, and starting a new daemon
with those parameters.