In IDEA configuration ignore resource dir declaration if the target is a source dir
As explained at https://github.com/gradle/gradle/issues/654 if a
directory is both declared as a source and a resource directory, then
the resource flag was added to it. This makes IDEA unusable as the
source editing is disabled for the contained source files. To overcome
that, this commit makes the resource declaration omitted if the target
is a source directory.