Allow testing with `java.system.class.path` and SLF4j
The core `Slf4jLoggingConfigurer` assumes that the Gradle implementation
of `org.slf4j.ILoggerFactory` is being used. This is not the case when
tests are run using a custom system classloader (by setting the
`java.system.class.loader` system property).

Previously, this would result in a `ClassCastException` when attempting
to configure the logger. This change makes `Slf4jLoggingConfigurer` lenient
in this case, opting not to configure the underlying logger rather than
fail.

Fixes #2657