Fix caching of component metadata rules in presence of `withDependencies`
This commit fixes several issues with component metadata rule `@CacheableRule`:

1. when using Gradle metadata, the original content hash wasn't set, leading
to incorrect keys in cache. The consequence is that we would fetch from cache
the wrong components.
2. there was a race condition between the component metadata serializer and
the graph resolution, which lead to rules being executed several times, and
actually multiple `DefaultConfigurationMetadata` created for the same components
3. a `ClassCastException` was visible as soon as one rule worked on dependencies
or dependency constraints, because we assume immutability, and that the rule
applicator was actually creating a mutable list
4. realized component metadata wasn't setting the `force` flag properly for
enforced platforms

This fixes #6978