Exclude application plugin from behaved plugins
With the ApplicationPluginIntegrationTest inheriting the tests from
WellBehavedPluginTest, it fails the "does not realize all possible
tasks" test. In order to make the test pass, the application plugin must
be included in the list of misbehaving plugins inside that test. This
confirms that the application plugin in its current state is not lazy.

Signed-off-by: Kevin Macksamie <kevin.macksamie@gmail.com>