Clear remaining edges when module is pending again
When a module moves back to pending - aka no hard edges left to it - we
were not removing the remaining constraint edges that could point to it.
We can simply clear these edges and not restart their origin as we know
the scope of the change is only about dropping constraints.

Fixes #7060