Add annotationProcessor Configuration to IDEA
Since we no longer want users to put their annotation processors
on the compile classpath, we need a solution that still makes them
run in IDEA. This fix makes them work just like before, i.e. still
putting them on the compile classpath in IDEA, while they are separated
when building with Gradle.

As a next step, we should provide the processor path as a tooling model
so that the IDEA team can configure the processor settings correctly.
Then we can remove this backward compatibility fix in Gradle 6.0