Do not use local `.m2` repository when running integration tests
Some integration tests expect that metadata is missing from `.m2`.
In practice we have something to avoid that `.m2` leaks on CI, but
nothing for local development. It appears it's pretty simple to
break integration tests because of the accidental presence of a
component in your local .m2 repository. This commit makes sure
that we isolate by default on all integration tests.