Try to execute unparseable test class files
Instead of failing the build when a test class file cannot be parsed,
e.g. if it's compiled for a JVM that is not yet supported by the
version of ASM that we use, we now assume it's a test class and pass
it to the testing framework which usually ignores classes that are not
tests.

Related issue: #7059