Fix SelectorState when removed from target module
When removing a `SelectorState` from its target module, it also needs to
be marked as _not_ resolved so that if it gets used again in the future
it will trigger proper selection again.
Doing this change uncovered a bug around conflict resolution. When a
module is forced to a different module, we need to remember this so that
any selector added after the force takes place can be marked as
overridden as well.

Fixes #7050