Implement retries with exponential backoff
This commit reworks the strategy used to blacklist repositories. In the case
an error occurs when trying to access a remote resource, if the error is not
a missing resource, we're going to retry twice before actually blacklisting.

Between each try, we're going to wait, and the wait is increasing between
each trial exponentially. There are two internal parameters which allow
tweaking the behavior:

- `org.gradle.internal.repository.max.retries` (default 3) is the number
of retries (initial included)
- `org.gradle.internal.repository.initial.backoff` is the initial time
before retrying, in milliseconds (default 125)

Fixes #4629