Rework where we check for mutation in DomainObjectSets only
This moves the special handling of "immutable" domain object collections into the
only implementation that actually uses it.

DefaultConfiguration uses this to prevent the addition/removal of dependencies after
dependency resolution has occurred, but this doesn't prevent all kinds of mutations.

The assertMutable(...) check was intended to check for _all_ kinds of mutation, but
this breaks the way DefaultConfiguration tries to prevent mutation (only on add/remove).

We should fix DefaultConfiguration to prevent all kinds of mutation, but that's a much bigger
problem.  This change restricts the specialness to just DomainObjectset.