Lower memory for the Gradle daemon by default
We have made lots of memory usage improvements to Gradle,
so 512m will be enough for small-medium sized projects.
The assumption here is that power users with large projects
are much more likely to tweak these settings than peole just
getting started.

Also this change limits the metaspace on Java8+, so that memory
leaks in plugins don't go unnoticed.