Reformat the release notes for Gradle 5.0
This changes how the release notes are laid out. It begins with
a very brief list of major features, thanks our contributors, and
then expands each of the features in detailed subsections.

This format is intended to break the very long preamble into the
detailed sections. We've found this format to increase engagement
in the newsletter. I also think this will increase the focus on
the community and encourage contribution, as well as eliminate the
difference between what's in the preamble and the detailed sections.

Finally, "Breaking Changes" has been moved to the upgrade guide and
replaced with a summary of breaking changes.
This is crucial because release notes are _not_ docs as they are
short-lived.