Add support for filtering repository content based on the module identifier
This commit allows "matching repositories to dependencies" by adding an API
on repositories to declare if we should look for a module or not. Currently
the API only supports querying by module identifier, and blacklisting. In
other words the API lets you tell when _not_ to look for a module.