Adds missing breaking changes to the upgrade guide
This change adds more breaking changes that were previously in the release notes but were absent from this upgrade guide.

Since the document was getting quite long in some sections, I moved some of the content.

I also added links to some of the notable libraries and their changelogs when there are version upgrades.