Fix classloader leaks in Groovy
We already had some workarounds for the ClassInfo
classloader leak introduced in Groovy 2.4. It is still
there in Groovy 2.5, but the class has slightly changed,
so our workaround needed some adjustments.

There was another yet undetected leak related to our use of
the GroovyShell in the ApiGroovyCompiler. Whenever the shell
is used, it would register a PreferenceListener and never
remove it, leaking the whole classloader that loaded this
GroovySystem. Removing this listener unfortunately required
doing yet another illegal reflective access, so hopefully
this can be fixed on the Groovy side in the future.