Fix metaspace leak in ShutdownHookActionRegister
This class added a shutdown hook and never removed it,
which created a classloader leak whenever it was used
in an embedded context, e.g. in the in-process Groovy
compiler.

The approach of having a single hook with multiple actions
seems like a premature optimization. Creating a Thread object
is not expensive, starting it is. The class now just serves
as a convenience for adding/removing shutdown hooks, but creates
one thread per hook, so it can be correctly removed again.