Limit metaspace used by the Gradle daemon
Our integration tests still need too much metaspace because
we never clean up buildSrc/script classloaders. This is a problem
in our tests since we create hundreds of projects, all with different
directories. For our end users this is probably not as bad, but we
should still fix it soon.