Rejig the user manual navigation
I've removed the link to the Using the Build Cache Guide as it wasn't clear
what the difference was with the Enabling and Configuring the Build Cache was.
To make up for its removal, the Guide is prominently linked from the user
manual chapter.

I've also reordered the Gradle API entries to improve the look by keeping the
Groovy DSL entries together and the Kotlin DSL ones.

Finally, I removed the Standard Gradle Plugins page because it overlaps with
the plugin reference page.