Use JavaParser to generate DSL metadata
We no longer have any Groovy source files that are part of the API.
Thus, support for parsing them is no longer necessary. The Java parser
that was used until now does not support Java 7 or 8 source code
features such as try-with-resources, lambdas, and method references.
Now, the JavaParser library is used which supports all of them.

Resolves #7566.