Show more edge information in dependencyInsight
The `dependencyInsight` report shows detailed information about a particular
node in the dependency graph, including all incoming edges to that node traced from
the root. Understand which incoming edge is responsible for influencing the selected
version is an important use case for `dependencyInsight`.

A previous change added and conflict resolution reasons that were present for a
node on an incoming edge. This PR brings an improvement over this previous behaviour:
with this change we show the requested version for each edge where it is different
from the actual version that was selected.