Add 'constraint' property to dependencies in ResolutionResult
It's important to be able to differentiate between hard dependencies and
dependency constraints when analysing the resolution result. This was done
by adding a simple boolean property to `DependencyResult`, indicating if
the source of the edge was a dependency or a constraint.

The new API is not yet used in any dependency reporting.