Do not attach children to constraints in `dependencies` report
Prior to this change, constraints and dependencies were treated identically in
the `dependencies` report. This meant that transitive dependencies would often
appear to be "brought in" by an imported BOM file or other source of dependency
constraints. This can be confusing to a user, who thinks of the constraint as
merely influencing the version of the dependency, not adding a new dependency.

This has been addressed by ensuring that transitive dependencies are attached to
a hard dependency in the dependencies report. Dependency constraint edges are
rendered non-transitively, even if they appear earlier in the dependency graph.

Fixes #5969
Fixes #4846