Support project path queries in dependencyInsight
Adds support for specifying a full project path to select a dependency
to view in the `dependencyInsight` report. This was working for many
cases due to matching on `ModuleVersionIdentifier` for the project.
With this change, it is now possible to use the full project path to
select a project dependency, even when the project is deeply nested.

Fixes #6857