Do not attempt to invoke included build for task that has already executed
We were only tracking the tasks that had been _requested_ for execution.
With this change, we also track the tasks that are executed transitively.
This will make composite builds work in cases where the included build
provides a jar to the classpath of an including build, and the classes
directory to the implementation configuration of the same build.

Fixes #7650