Allow plugins defined in included builds to be resolved using the `plugins { }` block.
The `plugins { }` block in a build script will now substitute a binary plugin with a plugin that has the requested plugin id and that is defined in an included build using the `java-gradle-plugin` plugin. Does not substitute plugins defined in the root build, or in the consuming included build, or in any source dependency. Does not substitute plugins that are not defined using the `java-gradle-plugin`.