ProgressLogEventGenerator gives NoSuchElementException #3465
Guards the progress logger against NPEs if an operation is not found in the local cache.
Various people reported for that particular code to crash. The source of the cache
inconsistency is not fixed with this commit, but at least does not let Gradle crash.
Most likely the entire code base in the area needs to be hardened against multi-threading.

Signed-off-by: Remo <remo.meier@adnovum.ch>