Use an identity to determine the workspace
There are still a few things missing for chained transforms. The
non-chained transforms should work fine already.

We now need to add the inputs to the UnitOfWork, since we will not only
have empty workspaces and we need to check the inputs for up-to-date
as well.

Currently, we do not re-execute transforms with the same identity at
the project level. Maybe we need to use the full inputs hash for
determining whether we need to re-execute (or fail) if a transformation
would be executed twice in the same workspace in the same build with
different inputs.