Fix attribute disambiguation ignoring attributes of different types
This is a hotfix for KT-28203, which is about the Kotlin2Js plugin
failing with 5.0 because of extra attributes disambiguation. What
happens is that we have 2 sets of attributes, but one is using the
coercing `Usage` attribute while the other is not. As a consequence,
the attribute is not removed from the list of extra attributes,
and we fail disambiguation.