Do not include an additional 'build cancelled' exception in the build result when there are some task failures to report.
Previously, this was sometimes happening and sometimes not, depending on whether the  worker thread that ran the failed task happened to clean up the queued tasks based on the failure before some other work thread happened to see the cancellation. Now, if a task happens to fail due to cancellation, the result the same as is as if the had of failed in any other way.

Also clean up the tapi tests to deal with the fact that this is not fixed in older Gradle versions.