Remove short circuiting cached execution of transforms
This commit removes the short circuit code path where we avoided executing a transform via the BuildOperationExecutor if we knew already that it would be served from the in-memory cache. This was a change to reduce the overhead of the executor, but this approach does not work well with more complex inputs used for artifact transforms.