Make JavaCompile result orthogonal to task outcome
While we currently only report rich results for successful `JavaCompile`
tasks, we might change that in the future. Therefore, this commit
renames `JavaCompileTaskSuccessResult` to
`JavaCompileTaskOperationResult` so we won't have to change the API
when we start reporting rich results for failed tasks as well.
There are only implementations that also implement `TaskSuccessResult`
and `InternalTaskSuccessResult`, respectively, though.