Fix for starting a single-use daemon when client has low memory
The previous logic was testing whether the daemon parameters had
a memory setting, but not whether that setting was the default or
whether it actually came from the user. The default settings are
ignored for the purpose of forking a single-use daemon, so they
also need to be ignored in the decision whether the current VM
has enough memory.