Snapshot primary inputs only once
That makes the immutable transformations faster, since we don't need
to ask the file system mirror for the same information twice - once
for calculating the identity and another time for snapshotting the
inputs.

For mutable transformations, we still need to snapshot the inputs when
we determine if they are up-to-date. So we now also snapshot when we
determine the identity.