Emit build operations around container callback executions (core and dependencyMgmt containers) (#7734)
* Decorate taskcontainer callbacks to track application id
* Decorate plugin container callbacks
* Decorate repositoryContainer callbacks
* Decoreate configurations and configuration.dependencies container callbacks
* Decorate artifactTypesDecorator callbacks
* Dont emit build ops for internal declared callbacks
* Provide usercode context in beforeResolve / afterResolve callbacks
* keep compatibility for nebula.dependency-recommender plugin
* Put domain collection callback build ops behind feature toggle
* Decorate Provider.configure() methods
* Simplify container callback filtering and decoration

Previously, we had three classes collaborating to achieve this but now this is inlined into effectively one. While this creates a more complex implementation, that is still rather simple, it avoids the more problematic complexity of a complicated relationship between the three implementations that also required extra state and details to be propagated through all of the collections.