Make display names and order reliable
Prior to this commit, the build operation's display name was used for a
plugin identifier's display name. That included the target project and
other irrelevant parts. Now, the plugin id, class name, or script file
name is used instead.

The order of plugin configuration results is now more reliable.
Previously, the order was reversed and very unintuitive. Now, we sort by
the unique first application id.