Remove short circuiting cached execution of transforms (#7823)
This commit removes the short circuit code path where we avoided executing a transform via the BuildOperationExecutor if we knew already that it would be served from the in-memory cache. This was a change to reduce the overhead of the executor, but this approach does not work well with more complex inputs used for artifact transforms.

We now don't schedule artifact transformations on the
`BuildOperationQueue`, since they are already scheduled as part of
the execution when declared as inputs. File transformations are not scheduled as part of
the execution, so we keep running them on the `BuildOperationQueue`.