Track task creation and execution times for listener/config callbacks
Project evaluation listener callbacks such as `afterEvaluate {}` and
configuration callbacks of domain object collections such as `all {}`
are now attributed to the plugin that registered them. Thus, the origin
plugins for tasks created in such callbacks is correctly determined and
their execution times are added to the plugin's total execution time
that is reported to TAPI progress listeners.