Create partial dependency graph for build dependencies
Previously when computing the graph for build dependencies, only
artifact information was collected. This commit changes that to also
collect the dependency graph and make it available during execution
graph building.

This limited approach works because any time the graph can become more
complex, through substitutions and others, it will be fully resolved.